Corrected float values and set all values as Vanila default.

+++ Options/ParkinghallOptions.cs
@@ -5,7 +5,7 @@ public class ParkinghallOptions
     public string Name { get; set; } = null!;
 
     public int GarageCapacity { get; set; }
-    public int Comfortfactor { get; set; }
+    public float Comfortfactor { get; set; }
     public int Electricityconsumption { get; set; }
     public int Waterconsumption { get; set; }
     public int GarbageAccumulation { get; set; }
